[SYCLomatic] Add help function for Math API migration.
