--- clang/test/dpct/test_api_level/Util/api_test3.cu
@@ -1,17 +1,0 @@
-// RUN: dpct --format-range=none  --usm-level=none  --use-custom-helper=api -out-root %T/Util/api_test3_out %s --cuda-include-path="%cuda-path/include" -- -x cuda --cuda-host-only
-// RUN: grep "IsCalled" %T/Util/api_test3_out/MainSourceFiles.yaml | wc -l > %T/Util/api_test3_out/count.txt
-// RUN: FileCheck --input-file %T/Util/api_test3_out/count.txt --match-full-lines %s
-// RUN: rm -rf %T/Util/api_test3_out
-
-// CHECK: 2
-// TEST_FEATURE: Util_fast_length
-
-__device__ void foo() {
-  int i;
-  double d;
-  d = norm(i, &d);
-}
-
-int main() {
-  return 0;
-}
